ai-ethics-risk-diagnosis/requirements.txt
# LangChain & LangGraph
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langgraph==0.2.45
langsmith==0.1.147

# Vector Store
faiss-cpu==1.9.0
chromadb==0.5.23

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Web Search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
tiktoken==0.8.0

# Visualization & Export
matplotlib==3.9.3
reportlab==4.2.5
pandas==2.2.3

# Jupyter (optional)
jupyter==1.1.1
ipykernel==6.29.5